requirements.txt
langchain==0.0.271
faiss-cpu==1.7.4
openai==0.27.9
unstructured==0.10.4
tabulate==0.9.0
pdf2image==1.16.3
pytesseract==0.3.10
tiktoken==0.4.0
python-dotenv==1.0.0
pypdf==3.15.2
